INA-Prototype/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
